chore: Move amqp example to own package
