--- springwolf-examples/springwolf-amqp-example/src/main/java/io/github/stavshamir/springwolf/example/SpringwolfExampleApplication.java
@@ -1,14 +1,0 @@
-package io.github.stavshamir.springwolf.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringwolfExampleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringwolfExampleApplication.class, args);
-	}
-
-}
-
