+++ springwolf-examples/springwolf-amqp-example/src/main/java/io/github/stavshamir/springwolf/example/amqp/SpringwolfExampleApplication.java
@@ -1,0 +1,14 @@
+package io.github.stavshamir.springwolf.example.amqp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringwolfExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringwolfExampleApplication.class, args);
+	}
+
+}
+
