+++ springwolf-examples/springwolf-amqp-example/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.rabbitmq.password=guest
 spring.application.name=Springwolf example project - Amqp
 
 springwolf.enabled=true
-springwolf.docket.base-package=io.github.stavshamir.springwolf.example
+springwolf.docket.base-package=io.github.stavshamir.springwolf.example.amqp
 springwolf.docket.info.title=${spring.application.name}
 springwolf.docket.info.version=1.0.0
 springwolf.docket.info.description=Springwolf example project to demonstrate springwolfs abilities
