--- springwolf-examples/springwolf-amqp-example/src/test/java/io/github/stavshamir/springwolf/example/SpringwolfExampleApplicationTest.java
@@ -1,20 +1,0 @@
-package io.github.stavshamir.springwolf.example;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-@SpringBootTest
-class SpringwolfExampleApplicationTest {
-
-    @Autowired
-    private ApplicationContext context;
-
-    @Test
-    void testContext() {
-        assertNotNull(context);
-    }
-}
