+++ springwolf-examples/springwolf-kafka-example/src/main/java/io/github/stavshamir/springwolf/example/amqp/configuration/CustomWebSecurityConfigurerAdapter.java
