--- springwolf-examples/springwolf-kafka-example/src/main/java/io/github/stavshamir/springwolf/example/configuration/KafkaConfiguration.java
@@ -1,13 +1,0 @@
-package io.github.stavshamir.springwolf.example.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-
-@Configuration
-@EnableKafka
-public class KafkaConfiguration {
-
-    public final static String PRODUCER_TOPIC = "example-producer-topic";
-
-    public final static String CONSUMER_TOPIC = "example-consumer-topic";
-}
