+++ springwolf-examples/springwolf-kafka-example/src/test/java/io/github/stavshamir/springwolf/example/ProducerIntegrationWithDockerTest.java
@@ -1,8 +1,8 @@
 package io.github.stavshamir.springwolf.example;
 
 import io.github.stavshamir.springwolf.SpringWolfKafkaConfigProperties;
-import io.github.stavshamir.springwolf.example.consumers.ExampleService;
-import io.github.stavshamir.springwolf.example.dtos.ExamplePayloadDto;
+import io.github.stavshamir.springwolf.example.amqp.consumers.ExampleService;
+import io.github.stavshamir.springwolf.example.amqp.dtos.ExamplePayloadDto;
 import io.github.stavshamir.springwolf.producer.SpringwolfKafkaProducer;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.KafkaAdminClient;
@@ -22,7 +22,7 @@ import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
 
-import static io.github.stavshamir.springwolf.example.dtos.ExamplePayloadDto.ExampleEnum.FOO1;
+import static io.github.stavshamir.springwolf.example.amqp.dtos.ExamplePayloadDto.ExampleEnum.FOO1;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
